+++ src/__mocks__/vue.js
@@ -1,4 +1,4 @@
-const vue = require.requireActual('vue')
+const vue = jest.requireActual('vue')
 vue.config.productionTip = false
 vue.config.devtools = false
 module.exports = vue