+++ CHANGELOG.md
@@ -22,6 +22,14 @@ Please document your changes in this format:
 
 ## [Unreleased]
 
+## [8.4.0] - 2020-04-25
+### Added
+- Pagination for pickups [#1982] @ang-zeyu
+- More translations
+
+### Fixed
+- Karrot icon on dev.karrot.world did not show @tiltec
+
 ## [8.3.0] - 2019-12-02
 ### Added
 - Offers (behind feature flag) [#1799] @nicksellen
@@ -515,8 +523,10 @@ After months of planning and programming, the first version of ~the foodsaving t
 [#1690]: https://github.com/yunity/karrot-frontend/issues/1690
 [#1691]: https://github.com/yunity/karrot-frontend/issues/1691
 [#1799]: https://github.com/yunity/karrot-frontend/issues/1799
+[#1982]: https://github.com/yunity/karrot-frontend/issues/1982
 
-[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v8.3.0...HEAD
+[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v8.4.0...HEAD
+[8.4.0]: https://github.com/yunity/karrot-frontend/compare/v8.3.0...v8.4.0
 [8.3.0]: https://github.com/yunity/karrot-frontend/compare/v8.2.0...v8.3.0
 [8.2.0]: https://github.com/yunity/karrot-frontend/compare/v8.1.2...v8.2.0
 [8.1.2]: https://github.com/yunity/karrot-frontend/compare/v8.1.1...v8.1.2